Move fileOperations.js to utils
