+++ src/controllers/delete/index.js
@@ -1,5 +1,5 @@
 const constants = require('../../constants');
-const fileOperation = require('../../fileOperation');
+const fileOperation = require('../../utils/fileOperation');
 
 const del = async (req, res) => {
   const { fileName } = req.params;
