+++ src/controllers/read/index.js
@@ -1,5 +1,5 @@
 const constants = require('../../constants');
-const fileOperation = require('../../fileOperation');
+const fileOperation = require('../../utils/fileOperation');
 
 const read = async (req, res) => {
   const { fileName } = req.params;
